Remove function indirection for common options
