+++ src/commands/database/create.mjs
@@ -2,11 +2,11 @@
 import { ServiceError } from "fauna";
 
 import { container } from "../../cli.mjs";
-import { validateDatabaseOrSecret } from "../../lib/command-helpers.mjs";
 import { CommandError } from "../../lib/errors.mjs";
 import { faunaToCommandError } from "../../lib/fauna.mjs";
 import { getSecret, retryInvalidCredsOnce } from "../../lib/fauna-client.mjs";
 import { colorize, Format } from "../../lib/formatting/colorize.mjs";
+import { validateDatabaseOrSecret } from "../../lib/middleware.mjs";
 
 async function runCreateQuery(secret, argv) {
   const { fql } = container.resolve("fauna");
