+++ src/commands/schema/schema.mjs
@@ -1,6 +1,11 @@
 //@ts-check
 
-import { validateDatabaseOrSecret } from "../../lib/command-helpers.mjs";
+import { validateDatabaseOrSecret } from "../../lib/middleware.mjs";
+import {
+  ACCOUNT_OPTIONS,
+  CORE_OPTIONS,
+  DATABASE_PATH_OPTIONS,
+} from "../../lib/options.mjs";
 import abandonCommand from "./abandon.mjs";
 import commitCommand from "./commit.mjs";
 import diffCommand from "./diff.mjs";
@@ -8,7 +13,7 @@ import pullCommand from "./pull.mjs";
 import pushCommand from "./push.mjs";
 import statusCommand from "./status.mjs";
 
-export const localSchemaOptions = {
+export const LOCAL_SCHEMA_OPTIONS = {
   "fsl-directory": {
     alias: ["directory", "dir"],
     type: "string",
@@ -20,6 +25,9 @@ export const localSchemaOptions = {
 
 function buildSchema(yargs) {
   return yargs
+    .options(ACCOUNT_OPTIONS)
+    .options(DATABASE_PATH_OPTIONS)
+    .options(CORE_OPTIONS)
     .command(abandonCommand)
     .command(commitCommand)
     .command(diffCommand)
