--- src/lib/misc.mjs
@@ -1,3 +1,0 @@
-export function isTTY() {
-  return process.stdout.isTTY;
-}
